cleanup of flutter test
