+++ packages/flutter/test/widgets/keep_alive_test.dart
@@ -350,7 +350,7 @@ void main() {
     await tester.pump();
     expect(tester.binding.renderView.toStringDeep(minLevel: DiagnosticLevel.info), equalsIgnoringHashCodes(
       'RenderView#00000\n'
-      ' │ debug mode enabled - ${Platform.operatingSystem}\n'
+      ' │ debug mode enabled - HostPlatform.${Platform.operatingSystem}\n'
       ' │ window size: Size(2400.0, 1800.0) (in physical pixels)\n'
       ' │ device pixel ratio: 3.0 (physical pixels per logical pixel)\n'
       ' │ configuration: Size(800.0, 600.0) at 3.0x (in logical pixels)\n'
