+++ packages/flutter_test/lib/src/platform.dart
@@ -2,21 +2,14 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-import 'dart:io';
-
-// Whether the current dart code is running in an environment that was compiled
-// to JavaScript.
-const bool _kIsCompiledToJavaScript = identical(0, 0.0);
+import 'package:flutter/foundation.dart';
 
 /// Whether the test is running on the Windows operating system.
 ///
 /// This does not include test compiled to JavaScript running in a browser on
 /// the Windows operating system.
 bool get isWindows {
-  if (_kIsCompiledToJavaScript) {
-    return false;
-  }
-  return Platform.isWindows;
+  return currentHostPlatform == HostPlatform.windows;
 }
 
 /// Whether the test is running on the macOS operating system.
@@ -24,10 +17,7 @@ bool get isWindows {
 /// This does not include test compiled to JavaScript running in a browser on
 /// the macOS operating system.
 bool get isMacOS {
-  if (_kIsCompiledToJavaScript) {
-    return false;
-  }
-  return Platform.isMacOS;
+  return currentHostPlatform == HostPlatform.macOS;
 }
 
 /// Whether the test is running on the Linux operating system.
@@ -35,13 +25,11 @@ bool get isMacOS {
 /// This does not include test compiled to JavaScript running in a browser on
 /// the Linux operating system.
 bool get isLinux {
-  if (_kIsCompiledToJavaScript) {
-    return false;
-  }
-  return Platform.isLinux;
+  return currentHostPlatform == HostPlatform.linux;
 }
 
 /// Whether the test is running in a web browser compiled to JavaScript.
 bool get isBrowser {
-  return _kIsCompiledToJavaScript;
+  return currentHostPlatform == HostPlatform.browser;
 }
+
